refactpr(tags): Pull out the edit modal into a hook and a component
